CI: Move DOSv2 ftime tests to file
